+++ database/seeders/DatabaseSeeder.php
@@ -13,5 +13,6 @@ final class DatabaseSeeder extends Seeder
         $this->call(CmsSeeder::class);
         $this->call(MemberSettingsSeeder::class);
         $this->call(AiSeeder::class);
+        $this->call(BillingSeeder::class);
     }
 }